qlog: move streaming code to streamer module
